Move `getWinEventLogInfo` to `IAccessibleHandler.utils` module to work around circular import
